Added Vaadin Directory Badge
Howdy, it’s Binh from Vaadin Marketing team. We’re soon to publish Vaadin Directory, which will contain tons of awesome web components. I think you have a great component and it would be awesome to feature it in our Directory. So yeah, just simply add this badge to your README.md and your component will be automatically published there in the future.